Cheery-pick changes from floodyberry.
The previous change is a merge from floodyberry. The important,
algorithmic change in it is that the multiplication by 19 can be
performed on one of the 64-bit inputs to the 64x64 multiplication. This
is much better than multipling the 128-bit result by 19.

The rest of the changes involve persuading the compiler to do a better
job.

In this change, I've taken some of floodyberry's changes only. This is
because I favour a different balence between speed and readability.

The only algorithmic change in this commit is new way of handing
subtractions in the field. This is thanks to an idea from Emilia Kasper.

If you want the fastest code possible, take the previous commit and
replace the fdifference_backwards with the version from this commit.