Replace multiplication in 32-bit freduce_coefficients with shifts
Bizarrely, when told to generate 32-bit code for multiplying an
int64_t by 0x400000l, gcc 4.6 (and probably earlier) generate code
that contains an imul and a mul instruction, when really it should
be able to do fine with shift instructions.  So just use explicit
shift operations instead.

Also, this patch use the same trick as in freduce_degree to make
the multiply-by-19 avoid 64-bit multiplication.

Together, these changes reduce the output of speed-curve25519-donna
from 1415us to 1118us on my core2 desktop.