Only do the first 2 steps of 2nd freduce_coefficients loop
See the inline comments for an explanation of the implications
here: after you've done the first pass through the coefficients,
carried output[10] forward to output[0], and reduced output[1] and
output[2], you're in a "nearly reduced" state where everything but
output[2] is reduced, and |output[2]| <= 2^26.  (That is, it is
over by at most one.)

Since the largest value we compute before doing a freduce_degree
freduce_coefficients operation is a product of two sums, it's easy
to show that even if output[2] is equal to +/- 2^26, the inputs to
freduce_coefficients will all be under 2^61, which meets our
requirements.

The only operation that requires its inputs in fully-reduced form
is fcontract.  Fortunately, running an extra freduce_coefficients
on a "nearly reduced" value produces a reduced value. This happens
because if the value is fully-reduced, no carry operations will
happen, and if it is reduced except for output[2]=+/-2^26,
output[2] will become 0 and subsequent carry values will be at
most +/-1, thus stopping with output[2] equal to either 1 or 0,
which is then fully reduced.

This patch improves speed-curve25519-donna by about 13-14% on my
core2 laptop building with GCC 4.0.