Fix a gcc warning in speed-curve25519.c
It's not kosher to format a uint64_t with %lu: long is quite often
less than 64 bits wide.  Fortunately, the value that we want to format
here is the number of usec per operation, which won't brush up against
LONG_MAX unless you've pessimized the code until it takes over an hour
per call.