Use swap_conditional to fix data-dependent timing in cmult
Branching based on secret inputs can leak them to an attacker.  Let's
not do that.  This patch uses the same trick as in
curve25519-donna-c64 to make cmult() run in constant time.

It comes at a price, of course; an 8-18% performance hit, depending on
where I measure.  Still, there's no point in doing crypto fast if we
can't do it securely.