Make fcontract almost time-invariant by replacing the inner loops
We replace the old

           while (input[i] < 0) {
             input[i] += 0x2000000;
             input[i + 1]--;
           }

with

           const limb mask = input[i]>>63;
           const limb carry = -((input[i] & mask)>>25);
           input[i] += carry << 25;
           input[i+1] -= carry;

This requires a compiler where >> does sign extension on negative
numbers, but that's true pretty much everyplace, right?

The "mask" value gives us a way to leave the inputs unchanged if
input[i] >= 0.  The "carry" value works out to the number of times we
would have run through the loop.

This patch actually speeds fcontract() up a little, but fcontract is
not enough of the overal curve25519-donna profile to matter.

There is a remaining data-dependency in fcontract in the outer "do { }
while (input[0] < 0)" loop.